Add cupsGetDevices API (still issues with streaming).
Revamp DEBUG_printf/puts stuff to support logging to a file, and to include
a timestamp with every message.

Update Server: header to report CUPS/1.4 by default.


git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@7460 7a7537e8-13f0-0310-91df-b6672ffda945

+++ cgi-bin/help-index.c
@@ -3,7 +3,7 @@
  *
  *   On-line help index routines for the Common UNIX Printing System (CUPS).
  *
- *   Copyright 2007 by Apple Inc.
+ *   Copyright 2007-2008 by Apple Inc.
  *   Copyright 1997-2007 by Easy Software Products.
  *
  *   These coded instructions, statements, and computer programs are the
@@ -500,7 +500,7 @@ helpSaveIndex(help_index_t *hi,		/* I - Index */
     else
     {
       if (cupsFilePrintf(fp, "%s %d " CUPS_LLFMT " " CUPS_LLFMT " \"%s\" \"%s\"\n",
-                         node->filename, node->mtime,
+                         node->filename, (int)node->mtime,
                          CUPS_LLCAST node->offset, CUPS_LLCAST node->length,
 			 node->section ? node->section : "", node->text) < 0)
         break;
